backend/requirements.txt
fastapi
gunicorn
uvicorn
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
bcrypt==3.2.0
passlib[bcrypt]
python-multipart
google-generativeai
google-genai
groq
azure-storage-blob
